% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dates.R
\name{check_dates}
\alias{check_dates}
\title{Check Dates}
\usage{
check_dates(from, to)
}
\arguments{
\item{from}{String specifying the start date. Format: YYYY MM DD (any separator)}

\item{to}{String specifying the end date. Format: YYYY MM DD (any separator)}
}
\value{
Side effect (raise error) if invalid dates
}
\description{
This helper function checks if two dates (from and to) are valid or not.
Dates are valid if:
}
\details{
Are formatted as YYYY/MM/DD
    From Date <= To Date
}
