% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Helper Function}
\usage{
fetch_data(fetch_method, symbol, from, to)
}
\arguments{
\item{fetch_method}{a function to load data}

\item{symbol}{a character vector}

\item{from}{character vector specifying the start date (ISO8601 format)}

\item{to}{character vector specifying the end date (ISO8601 format)}
}
\value{
dataframe
}
\description{
Helper Function
}
