% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CAQV.R
\name{get_CAQV}
\alias{get_CAQV}
\title{Get Close Price, Adjusted Price, Quantity & Volume}
\usage{
get_CAQV(stock_name, from, to)
}
\arguments{
\item{stock_name}{a character vector of length one. Use \code{listed_symbols(type = 'stock')}
to list all valid stocks names listed in CSE.}

\item{from}{a character vector of length one matching the format 'yyyy-mm-dd'}

\item{to}{a character vector of length one matching the format 'yyyy-mm-dd'}
}
\value{
a dataframe.
}
\description{
Given a valid stock name listed in CSE, this function returns the corresponding financial
data including the ADJUSTED PRICES for the specified period (\code{from} until \code{to}).
The returned dataframe includes Date, Close, Adjusted, Quantity and Volume.
}
\note{
The difference between \code{to} and \code{from} dates should be less than or
equal 3 years.
}
\examples{
\dontrun{get_CAQV('COSUMAR', '2014/01/01', '2016/12/31')}
\dontrun{get_CAQV('ATLANTA', '2017/01/01', '2017/06/30')}

}
