% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OHLCap.R
\name{get_OHLCap}
\alias{get_OHLCap}
\title{Get Open, High, Low and Capitalisation}
\usage{
get_OHLCap(stock_name, from, to)
}
\arguments{
\item{stock_name}{a character vector of length one. Use \code{listed_symbols(type = "stock")}
to list all valid stocks names listed in CFM.}

\item{from}{a character vector of length one matching the format 'yyyy-mm-dd'}

\item{to}{a character vector of length one matching the format 'yyyy-mm-dd'}
}
\value{
a dataframe
}
\description{
Given a valid stock name listed in CSE, this function returns the corresponding financial
data for the specified period (\code{start_date} until \code{end_date}). The returned
dataframe includes Date, Open, High, Low and Capitalisation.
}
\note{
The difference between \code{to} and \code{from} dates should be less than or
equal 3 years.
}
\examples{
\dontrun{get_OHLCap('COSUMAR', '2014/01/01', '2016/12/31')}
\dontrun{get_OHLCap('ATLANTA', '2017/01/01', '2017/06/30')}


}
