% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OHLCAQVCap.R
\name{get_OHLCAQVCap}
\alias{get_OHLCAQVCap}
\title{Combine \code{get_CAQV} and \code{get_OHLCap}}
\usage{
get_OHLCAQVCap(symbol, from, to)
}
\arguments{
\item{symbol}{a character vector}

\item{from}{a character vector}

\item{to}{a character vector}
}
\value{
a dataframe
}
\description{
Combine \code{get_CAQV} and \code{get_OHLCap}
}
