% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbol.R
\name{get_symbol}
\alias{get_symbol}
\title{Load and Manage Historical Data from Casablanca Stock Exchange}
\usage{
get_symbol(symbol, from, to, type, format.as.xts = TRUE)
}
\arguments{
\item{symbol}{a character vector specifying the name of the symbol to be loaded. Use \code{listed_symbols} function to get all available symbol names.}

\item{from}{a character vector specifying the start date. Should be formatted according to ISO8601 (\code{"yyyy-mm-dd"})}

\item{to}{a character vector specifying the end date. Should be formatted according to ISO8601 (\code{"yyyy-mm-dd"})}

\item{type}{a character vector specifying the symbol type. Current supported types are "stock" and "index"}

\item{format.as.xts}{a boolean value. Set to \code{FALSE} in order to return a dataframe instead of an xts object}
}
\value{
an xts object or a dataframe
}
\description{
Function to load and manage historical data of a Symbol listed in Casablanca Stock Exchange.
Current available symbol types are "stock" and "index".
For stocks, the returned object includes: Date, Open, High, Low, Close, Adjusted, Quantity, Volume, and Capitalisation.
For indexes, the returned object includes: Date and Value columns.
}
\examples{
\dontrun{get_symbol('COSUMAR', '2017-01-01', '2017-12-31')}
\dontrun{get_symbol('COSUMAR', '2017-01-01', '2017-06-30', format.as.xts = TRUE)}
\dontrun{get_symbol('MASI', '2017-01-01', '2017-12-31', type = 'index')}


}
