% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes_metadata.R
\name{indexes_metadata}
\alias{indexes_metadata}
\title{Get Indexes Codes and Associated HTML Names (Description)}
\usage{
indexes_metadata(request = NULL, with_description = TRUE)
}
\arguments{
\item{request}{The object returned by an HTTP GET request. \code{NULL} by default.}

\item{with_description}{Boolean vector of length one indicating if the description
(HTML names) should be returned or not. \code{TRUE} by default.}
}
\value{
A character vector containing all available indexes codes.
If \code{with_description} is true, a data frame with INDEX_CODE and DESCRIPTION
columns is returned.
}
\description{
This helper function extracts the names and HTML values (options values) of all
available indexes in Casablanca Stock Exchange.
}
\examples{
\dontrun{indexes_metadata()}
\dontrun{indexes_metadata(with_description = FALSE)}
\dontrun{indexes_metadata(get_request, FALSE)}


}
