% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{INDEXES_XPATH}
\alias{INDEXES_XPATH}
\title{Where To Get The List Of All Available Indexes}
\format{An object of class \code{character} of length 1.}
\usage{
INDEXES_XPATH
}
\description{
Where To Get The List Of All Available Indexes
}
\keyword{datasets}
