% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipo_date.R
\name{ipo_date}
\alias{ipo_date}
\title{Get IPO Date of a Stock}
\usage{
ipo_date(symbol)
}
\arguments{
\item{symbol}{character string. Use \code{listed_symbols(type = "stock")}
to get all valid stock names in CSE}
}
\value{
Date
}
\description{
Get IPO Date of a Stock
}
\examples{
\dontrun{ipo_date('SNEP')}

}
