% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{IPO_DATE_XPATH}
\alias{IPO_DATE_XPATH}
\title{IPO Date XPATH}
\format{
An object of class \code{character} of length 1.
}
\usage{
IPO_DATE_XPATH
}
\description{
IPO Date XPATH
}
\keyword{datasets}
