% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{KEY_INDICATORS_BASE_URL}
\alias{KEY_INDICATORS_BASE_URL}
\title{Base URL to Get a Stock Key Indicators
Should be Appended with the Stock Code Value}
\format{An object of class \code{character} of length 1.}
\usage{
KEY_INDICATORS_BASE_URL
}
\description{
Base URL to Get a Stock Key Indicators
Should be Appended with the Stock Code Value
}
\keyword{datasets}
