% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{LISTED_SEC_XPATH}
\alias{LISTED_SEC_XPATH}
\title{Where To Get The List Of All Listed Stocks}
\format{
An object of class \code{character} of length 1.
}
\usage{
LISTED_SEC_XPATH
}
\description{
Where To Get The List Of All Listed Stocks
}
\keyword{datasets}
