% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listed_symbols.R
\name{listed_symbols}
\alias{listed_symbols}
\title{Get all Available Symbols in Casablanca Stock Exchange}
\usage{
listed_symbols(type = "all")
}
\arguments{
\item{type}{one of "stock", "index" or "all"}
}
\value{
if \code{type == "all"}: list with two columns (INDEXES, STOCKS)

if \code{type == "index"}: dataframe (index code and description)

if \code{type == "stock"}: character vector
}
\description{
This function returns a list of all stocks and indexes currently available in CSE.
The argument \code{type} controls the symbol type: stock, index or all (for both).
}
\examples{
\dontrun{listed_symbols()}
\dontrun{listed_symbols()$INDEXES}
\dontrun{listed_symbols()$STOCKS}
\dontrun{listed_symbols(type = 'stock')}
\dontrun{listed_symbols(type = 'index')}

}
