% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_metadata.R
\name{stocks_metadata}
\alias{stocks_metadata}
\title{Get Stocks Names and Associated HTML Values}
\usage{
stocks_metadata(request = NULL, with_values = TRUE)
}
\arguments{
\item{request}{The object returned by an HTTP GET request. \code{NULL} by default.}

\item{with_values}{Boolean vector of length one indicating if the HTML values should be
returned or not. \code{TRUE} by default.}
}
\value{
A character vector containing all listed stocks. If \code{with_value} is true, a
named vector of HTML values is returned.
}
\description{
This helper function extracts names and HTML values (options values) of all
listed stocks in Casablanca Stock Exchange.
}
\examples{
\dontrun{stocks_metadata()}
\dontrun{stocks_metadata(with_values = FALSE)}
\dontrun{stocks_metadata(get_request, FALSE)}

}
