% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{URL_INDEXES}
\alias{URL_INDEXES}
\title{URL For Indexes Data}
\format{
An object of class \code{character} of length 1.
}
\usage{
URL_INDEXES
}
\description{
URL For Indexes Data
}
\keyword{datasets}
