% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VIEW_STATE_XPATH}
\alias{VIEW_STATE_XPATH}
\title{VIEW_STATE XPath}
\format{An object of class \code{character} of length 1.}
\usage{
VIEW_STATE_XPATH
}
\description{
VIEW_STATE XPath
}
\keyword{datasets}
